+++ examples/react-typescript/src/core/app.ts
@@ -1,6 +1,0 @@
-import React from "react";
-import Pulse from "pulse-framework";
-
-export const App = new Pulse({
-	framework: React,
-});