+++ examples/react-typescript/src/core/controllers/accounts/index.ts
@@ -1,1 +1,0 @@
-export { accounts } from "./account.controller";