+++ examples/react-typescript/src/core/index.ts
@@ -1,3 +1,0 @@
-import { core } from './core';
-
-export default core;