+++ lib/collection/index.ts
@@ -1,3 +1,0 @@
-import Collection from './collection';
-
-export default Collection;