+++ lib/helpers/index.ts
@@ -1,8 +1,0 @@
-import * as handy from './handy';
-
-import { deepmerge } from './deepmerge';
-
-export default {
-  deepmerge,
-  ...handy
-};