+++ packages/pulse-core/lib/index.ts
@@ -1,0 +1,5 @@
+export * from './internal';
+
+import { Pulse } from './internal';
+
+export default Pulse;